upstream mega_downloader
thanks to @adek maulana

+++ userbot/modules/mega_downloader.py
@@ -18,15 +18,17 @@
 #  OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
 #  ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-from subprocess import PIPE, Popen
+from asyncio import create_subprocess_shell as asyncSubprocess
+from asyncio.subprocess import PIPE as asyncPIPE
 
+import asyncio
 import re
 import json
 import os
-import time
+import multiprocessing
+import errno
 
 from pySmartDL import SmartDL
-from os.path import exists
 from urllib.error import HTTPError
 
 from userbot import CMD_HELP, LOGS
@@ -35,19 +37,17 @@ from userbot.modules.upload_download import humanbytes
 
 
 async def subprocess_run(cmd, megadl):
-    subproc = Popen(cmd, stdout=PIPE, stderr=PIPE,
-                    shell=True, universal_newlines=True,
-                    executable="bash")
-    talk = subproc.communicate()
+    subproc = await asyncSubprocess(cmd, stdout=asyncPIPE, stderr=asyncPIPE)
+    stdout, stderr = await subproc.communicate()
     exitCode = subproc.returncode
     if exitCode != 0:
         await megadl.edit(
-            '```An error was detected while running the subprocess:\n'
-            f'exit code: {exitCode}\n'
-            f'stdout: {talk[0]}\n'
-            f'stderr: {talk[1]}```')
+            '**An error was detected while running subprocess**\n'
+            f'```exit code: {exitCode}\n'
+            f'stdout: {stdout.decode().strip()}\n'
+            f'stderr: {stderr.decode().strip()}```')
         return exitCode
-    return talk
+    return stdout, stderr
 
 
 @register(outgoing=True, pattern=r"^.mega(?: |$)(.*)")
@@ -70,7 +70,7 @@ async def mega_downloader(megadl):
     cmd = f'bin/megadown -q -m {link}'
     result = await subprocess_run(cmd, megadl)
     try:
-        data = json.loads(result[0])
+        data = json.loads(result[0].decode().strip())
     except json.JSONDecodeError:
         await megadl.edit("`Error: Can't extract the link`\n")
         return
@@ -102,27 +102,34 @@ async def mega_downloader(megadl):
         estimated_total_time = downloader.get_eta(human=True)
         try:
             current_message = (
-                f"**{status}**..."
-                f"\nFile Name: `{file_name}`\n"
-                f"\n{progress} `{percentage}%`"
+                "File Name:"
+                f"\n`{file_name}`\n\n"
+                "Status:"
+                f"\n**{status}** | {progress} `{percentage}%`"
                 f"\n{humanbytes(downloaded)} of {humanbytes(total_length)}"
                 f" @ {speed}"
                 f"\nETA: {estimated_total_time}"
             )
-            if status == "Downloading":
+            if display_message != current_message:
                 await megadl.edit(current_message)
-                time.sleep(0.2)
-            elif status == "Combining":
-                if display_message != current_message:
-                    await megadl.edit(current_message)
-                    display_message = current_message
+                await asyncio.sleep(0.2)
+                display_message = current_message
         except Exception:
             pass
+        finally:
+            if status == "Combining":
+                await asyncio.sleep(float(downloader.get_eta()))
     if downloader.isSuccessful():
         download_time = downloader.get_dl_time(human=True)
-        if exists(temp_file_name):
-            await decrypt_file(
-                file_name, temp_file_name, hex_key, hex_raw_key, megadl)
+        try:
+            P = multiprocessing.Process(target=await decrypt_file(
+                file_name, temp_file_name, hex_key, hex_raw_key, megadl), name="Decrypt_File")
+            P.start()
+            P.join()
+        except FileNotFoundError as e:
+            await megadl.edit(str(e))
+            return
+        else:
             await megadl.edit(f"`{file_name}`\n\n"
                               "Successfully downloaded\n"
                               f"Download took: {download_time}")
@@ -135,11 +142,13 @@ async def mega_downloader(megadl):
 
 async def decrypt_file(file_name, temp_file_name,
                        hex_key, hex_raw_key, megadl):
-    await megadl.edit("Decrypting file...")
     cmd = ("cat '{}' | openssl enc -d -aes-128-ctr -K {} -iv {} > '{}'"
            .format(temp_file_name, hex_key, hex_raw_key, file_name))
-    await subprocess_run(cmd, megadl)
-    os.remove(temp_file_name)
+    if await subprocess_run(cmd, megadl):
+        os.remove(temp_file_name)
+    else:
+        raise FileNotFoundError(
+            errno.ENOENT, os.strerror(errno.ENOENT), file_name)
     return
 
 
@@ -148,5 +157,5 @@ CMD_HELP.update({
     ".mega <mega url>\n"
     "Usage: Reply to a mega link or paste your mega link to\n"
     "download the file into your userbot server\n\n"
-    "Only support for *FILE* only.\n"
+    "Only support for *FILE* only."
 })
